requirements.txt
flask==3.1.2
numpy==1.22.4
Pillow==10.2.0

torch==2.8.0
torchvision==0.23.0